fix: resovle exception handler sequence
