+++ backend/project-seckill-api/src/main/java/cn/edu/ntu/seckill/handler/BusinessExceptionHandler.java
@@ -3,6 +3,7 @@ package cn.edu.ntu.seckill.handler;
 import cn.edu.ntu.model.ErrorMessageEnum;
 import cn.edu.ntu.model.ErrorResponse;
 import cn.edu.ntu.seckill.exception.UserException;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -16,6 +17,7 @@ import javax.servlet.http.HttpServletRequest;
  * @create 2020-07-21 23:48 <br>
  * @project project-seckill <br>
  */
+@Order(100)
 @ResponseBody
 @ControllerAdvice
 public class BusinessExceptionHandler extends BaseExceptionHandler {
